jlログファイルを格納するディレクトリを変更
