+++ jsonl/__init__.py
@@ -1,0 +1,3 @@
+# coding:utf-8
+from __future__ import unicode_literals
+
