+++ jsonl/rw.py
@@ -5,8 +5,8 @@ import datetime
 import json
 
 
-def jlfile_path(date, pref=''):
-    return "./log/{0}{1:%Y%m%d}.jl".format(pref, date)
+def jsonl_file_path(date, pref=''):
+    return "./resources/jsonl/{0}{1:%Y%m%d}.jsonl".format(pref, date)
 
 
 class Writer(object):
@@ -38,7 +38,7 @@ class Writer(object):
             cls=self.Encoder
         )
 
-    def write_jsonlines(self, list_of_serializable, delim="\n", file_close=False):
+    def write(self, list_of_serializable, delim="\n", file_close=False):
         """:type list_of_serializable: list[dict|list]"""
         self.file.writelines([
             (self.to_json(d) + delim).encode('utf-8')
