+++ resources/jsonl/.gitignore
@@ -1,0 +1,1 @@
+# http://jsonlines.org/
